moved settings module
