+++ services/director-v2/src/simcore_service_director_v2/api/dependencies/dynamic_sidecar.py
@@ -1,6 +1,8 @@
 from fastapi import Request
 
-from ...core.dynamic_sidecar_settings import DynamicSidecarSettings
+from ...core.dynamic_services_settings.dynamic_sidecar_settings import (
+    DynamicSidecarSettings,
+)
 from ...modules.dynamic_sidecar.scheduler import DynamicSidecarsScheduler
