+++ services/director-v2/src/simcore_service_director_v2/core/dynamic_services_settings/dynamic_sidecar_settings.py
@@ -11,9 +11,9 @@ from settings_library.r_clone import RCloneSettings as SettingsLibraryRCloneSett
 from settings_library.utils_logging import MixinLoggingSettings
 from settings_library.utils_service import DEFAULT_FASTAPI_PORT
 
-from ..constants import DYNAMIC_SIDECAR_DOCKER_IMAGE_RE
-from .dynamic_services_settings.egress_proxy import EgressProxySettings
-from .dynamic_services_settings.proxy import DynamicSidecarProxySettings
+from ...constants import DYNAMIC_SIDECAR_DOCKER_IMAGE_RE
+from .egress_proxy import EgressProxySettings
+from .proxy import DynamicSidecarProxySettings
 
 _logger = logging.getLogger(__name__)
