+++ services/director-v2/src/simcore_service_director_v2/modules/dynamic_sidecar/docker_service_specs/proxy.py
@@ -9,8 +9,10 @@ from models_library.services_resources import (
 )
 from pydantic import ByteSize
 
+from ....core.dynamic_services_settings.dynamic_sidecar_settings import (
+    DynamicSidecarSettings,
+)
 from ....core.dynamic_services_settings.proxy import DynamicSidecarProxySettings
-from ....core.dynamic_sidecar_settings import DynamicSidecarSettings
 from ....models.dynamic_services_scheduler import SchedulerData
 from ._constants import DOCKER_CONTAINER_SPEC_RESTART_POLICY_DEFAULTS
