+++ services/director-v2/src/simcore_service_director_v2/modules/dynamic_sidecar/docker_service_specs/volume_remover.py
@@ -17,7 +17,9 @@ from models_library.users import UserID
 from pydantic import parse_obj_as
 
 from ....constants import DYNAMIC_VOLUME_REMOVER_PREFIX
-from ....core.dynamic_sidecar_settings import DynamicSidecarSettings
+from ....core.dynamic_services_settings.dynamic_sidecar_settings import (
+    DynamicSidecarSettings,
+)
 
 
 class DockerVersion(str):
