+++ services/director-v2/src/simcore_service_director_v2/modules/dynamic_sidecar/scheduler/_core/_observer.py
@@ -9,10 +9,12 @@ from fastapi import FastAPI
 from servicelib.error_codes import create_error_code
 
 from .....core.dynamic_services_settings import DynamicServicesSettings
+from .....core.dynamic_services_settings.dynamic_sidecar_settings import (
+    DynamicSidecarSettings,
+)
 from .....core.dynamic_services_settings.scheduler import (
     DynamicServicesSchedulerSettings,
 )
-from .....core.dynamic_sidecar_settings import DynamicSidecarSettings
 from .....models.dynamic_services_scheduler import (
     DynamicSidecarStatus,
     SchedulerData,
