+++ services/director-v2/src/simcore_service_director_v2/modules/dynamic_sidecar/scheduler/_core/_scheduler_utils.py
@@ -10,10 +10,12 @@ from models_library.projects_nodes_io import NodeID
 from models_library.services_enums import ServiceBootType, ServiceState
 from servicelib.fastapi.long_running_tasks.client import ProgressCallback
 
+from .....core.dynamic_services_settings.dynamic_sidecar_settings import (
+    DynamicSidecarSettings,
+)
 from .....core.dynamic_services_settings.scheduler import (
     DynamicServicesSchedulerSettings,
 )
-from .....core.dynamic_sidecar_settings import DynamicSidecarSettings
 from .....models.dynamic_services_scheduler import DynamicSidecarStatus, SchedulerData
 from ...api_client import SidecarsClient, get_sidecars_client
 from ...docker_api import (
