+++ services/director-v2/tests/integration/02/utils.py
@@ -31,7 +31,7 @@ from simcore_service_director_v2.constants import (
     DYNAMIC_PROXY_SERVICE_PREFIX,
     DYNAMIC_SIDECAR_SERVICE_PREFIX,
 )
-from simcore_service_director_v2.core.dynamic_sidecar_settings import (
+from simcore_service_director_v2.core.dynamic_services_settings.dynamic_sidecar_settings import (
     DynamicSidecarSettings,
 )
 from simcore_service_director_v2.models.dynamic_services_scheduler import SchedulerData
