+++ services/director-v2/tests/unit/with_dbs/test_modules_dynamic_sidecar_docker_api.py
@@ -29,7 +29,7 @@ from simcore_service_director_v2.constants import (
     DYNAMIC_SIDECAR_SERVICE_PREFIX,
     DYNAMIC_VOLUME_REMOVER_PREFIX,
 )
-from simcore_service_director_v2.core.dynamic_sidecar_settings import (
+from simcore_service_director_v2.core.dynamic_services_settings.dynamic_sidecar_settings import (
     DynamicSidecarSettings,
 )
 from simcore_service_director_v2.models.dynamic_services_scheduler import (
