+++ services/director-v2/tests/unit/with_dbs/test_modules_dynamic_sidecar_docker_service_specs.py
@@ -24,7 +24,7 @@ from pytest import MonkeyPatch
 from pytest_simcore.helpers.typing_env import EnvVarsDict
 from pytest_simcore.helpers.utils_envs import setenvs_from_dict
 from servicelib.json_serialization import json_dumps
-from simcore_service_director_v2.core.dynamic_sidecar_settings import (
+from simcore_service_director_v2.core.dynamic_services_settings.dynamic_sidecar_settings import (
     DynamicSidecarSettings,
 )
 from simcore_service_director_v2.models.dynamic_services_scheduler import SchedulerData
